requirements.txt
django==4.2
djangorestframework==3.15
psycopg2-binary==2.9.9
pgvector==0.2.5
redis==5.0.4
celery==5.3.6
python-dotenv==1.0.1
langchain==0.2.0
langchain-text-splitters==0.2.0
sentence-transformers==2.7.0
